feat: add SBI processor
